requirements.txt
customtkinter>=5.2.0
openpyxl>=3.1.0
reportlab>=4.0.0
matplotlib
Pillow
tkfontawesome>=0.2.0
pyinstaller>=6.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
